+++ camera.go
@@ -3,9 +3,13 @@ package main
 type Camera struct {
 	position  Vec3
 	direction Vec3
-	up        Vec3
+	velocity  Vec3
+	yaw       float64
 }
 
-func NewCamera(position, direction, up Vec3) Camera {
-	return Camera{position: position, direction: direction, up: up}
+func NewCamera(position, direction Vec3) Camera {
+	return Camera{
+		position:  position,
+		direction: direction,
+	}
 }